remove recursive style codegen (full switch to flat codegen)
