--- codegen/tmpl/main.tmpl
@@ -1,14 +1,0 @@
-{{ define "main" }}// Code generated by dynamic-ssz. DO NOT EDIT.
-// Hash: {{ .TypesHash }}
-// Version: {{ .Version }} (https://github.com/pk910/dynamic-ssz)
-package {{ .PackageName }}
-
-import (
-	{{- range .Imports }}
-	{{ if .Alias }}{{ .Alias }} {{ else }}{{ end }}"{{ .Path }}"
-	{{- end }}
-)
-
-var _ = sszutils.ErrListTooBig
-{{ .Code }}
-{{- end }}
